FISH-10240 PR review changes
